pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.20",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.19",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.6",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "newsapi-python>=0.2.7",
    "python-dotenv>=1.1.1",
]
